perf: update drawer size and ticket detail
